services/worker/requirements.txt
celery==5.3.6
redis==5.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0